Increased button debounce time and other minor fixes
The button debounce time doesn't need to be so short and a longer time should help problematic buttons.

Added amber and yellow to the colours and changed the no IR seen colour to amber (orange and salmon look similar).

Removed the DTR macro since the RP2040 has the same dtr() method.