Hotfix for events not firing when not-trigger buttons are pressed while firing
Turns out, buttons.debounced is a bitmask and not a byte array, so we really SHOULD be reading these one bit at a time. Rightmost bit correlates to 1, which is the trigger.

Also added an exception when start or select are pressed while the trigger is pulled (in the case of games like Crisis Zone PS2 that exits with Start + Trigger), it just feels cleaner. User can always remove the latter two bitreads in the comparison if they want.