RUMBLE_FF prep work
Because I was bored, set things up for the eventual rumble-based surrogate force feedback system.
Things that explicitly use Solenoid or Rumble are now processed and inserted/removed via compiler flags, and added placeholders for RUMBLE_FF to be filled in later (which will be incompatible with normal rumble, so the user will have to decide which they'll have). No other functional changes made here.

You ever think about the definition of insanity?