EEPROM.begin is only for RP2040
Make sure it's only enabled if the EEPROM-toting device is an RP2040 specifically, since it's for the emulated eep and not a real one. Building for ATmega had a fit over this, so making sure.

Also forgot one little LED function that wasn't being hidden behind an LED_ENABLE flag. whoops.

Turns out, building this project for Pro Micro *can* work... if you disable solenoid and rumble motor support, and aren't using hw switches, it'll fit into the allotted ~32K of project storage... barely (28638 used out of 28672!). But for how much longer? At that point, we're only a few degrees shy of just being Prow's fork with a keyboard debounce bugfix.