The RP2040 dual cores update
Since the Cortex M0+ that the RP2040 uses has two cores, why not use them?

This splits the input processing load off from the main thread (core0, which mainly handles camera updates and tracking now) into its own dedicated loop (core1). Just this on its own noticeably improves camera update performance, and (possibly?) reduces trigger/buttons lag.

For some reason, though, the second core brings the performance to a screeching halt if it does any *other* bitreads that aren't the trigger. No, the performance of the trigger itself isn't affected in any way - just the keyboard buttons. So we tentatively rely on the old method with the LightgunButtons library - which isn't a problem, because it doesn't seem to have the same debouncing issues as single threaded? Whatever, works for us ig!

Only applies to an RP2040 board, obvi; if you use one, but want to roll back to single threaded performance mode or compare the performance yourself, unset `DUAL_CORE`.