Optimize trigger functions
When I moved the trigger detection stuff into its own function, I left it as one giant function so the sketch was jumping to the start of it every cycle, which could cause needless checks and potentially (a bit) more overhead. It's also just not very clean.
This splits it off to two functions; one if trigger is pulled, the other when not, which makes it behave more closely to before.