The fix everything update
Because... WHOO boy did I have a lot to explain for.

Like the fact that throwing in all those button checks destroyed performance on most (single core) boards! We not only brought the special input handling to its own function (more on that), but fixed it so it wasn't spamming return signals every frame, which artificially noticeably lowered camera tracking rate. Now it should be as smooth as the original fork, with less (if any) difference between single and dual-core devices.

The main thing though is the buttons handling. Notes in the code, but to the point, it's not LightgunButtons' fault; it's the keyboard (and ONLY the keyboard btw) inputs getting "lost in traffic" alongside camera updates. We now handle those in a slightly hacky, but infinitely more consistent manner without breaking performance, at the cost of *potential* jitter while pressing keyboard buttons in motion. The values chosen have been picked to mitigate this pause as much as possible.

And with that, the dual core update is rendered kinda moot? Lol. But it handles button inputs anyways and has a chance of reducing latency, so it's maintained and shouldn't cause issues either way.

FUN FACT: didjuknow that the reason the original dual core input handling caused the camera to die is because the second core is so fast that it completely jammed the USB interface? Yeah, it's that stupid quick. Even with the delay-based timing, there's no noticeable pause, so still a benefit (if a theoretical one) to having it.