The (probably incredibly broken) MAMEHooker update
Yes, it's finally, kinda, sorta, maybe here.

This introduces (basic, unfinished) support for MAMEHooker's outputs, handing off control if the right serial byte is detected. We currently only process S (start), E (end), M1x2 (offscreen button mode), and F0 & F1 (solenoid/rumble feedback), with plans to implement F2/3/4 for controlling the board's preinstalled LED if available.

This can be disabled if desired, but it will only come into effect if the gun has the S start command sent to it - and if it gets stuck somehow (not sure if this is a thing), a ten second timeout will forcibly disable serial handoff and let the gun handle force feedback on its own again.

I have... NO IDEA if this actually works - but all the listed supported modes did react to the Arduino IDE how I expected? Blame the lack of clean documentation for this, as the only semi-complete resource I was given needed to be translated from Spanish. :/