MAMEHooker fixes
Implemented serial handoff specific offscreen button mode, so it won't interfere with the normal gun mode's offButt mode.

Fixed how we interpreted the M1 bit - it was blindly setting offscreen button mode anytime the M1 command was seen. Now we actually parse if it's asking for "offscreen button press" mode specifically, and only then do we flip the serial offButt mode on.

Throughout this project's life, I've noticed that the RP2040 would randomly crash when it was receiving or sending a huge influx of serial data in a short time. To mitigate this, I've implemented a short 1ms pause at the end of each main runmode loop (single or dual core, doesn't matter). Hopefully this is the root cause, as it seems to occur at complete random.

That said, MAMEhook support *should* be pretty much done from the feedback handling side. All that's left, from my personal want list, is parsing and processing LEDs...