(Partial) Serial LED support
Adafruit boards' builtin DotStar or NeoPixel (either or, if LED_ENABLE is set) are now used as part of the serial handover mode.

As such, F2-4 are now (partially) implemented.

When a start command is received, the LED will default to a mid-intensity white. If any further commands are sent down either the F2 (R), F3 (G), or F4 (B) paths, the LED will change accordingly.
LED states only change when either an explicit command has been sent (mostly so the initial on color stays on), or when an E (exit) command has been sent, to which the LED will turn off and set all the serial LED values to 0 (from which it'll either stay off, or glow as according to UpdateLastSeen();).

This is only partial support, as we currently only parse Fx.0/1 (off/on). The Fx.2 (pulse) command bit will simply be discarded as well as its associated value.