Added solenoid pulses
The solenoid was originally using digitalWrites regardless of a normal on or pulse on signal. Now pulse on signals use PWM to (somewhat?) smoothly pull and release the armature.

This has the small side effect that the PWM activity *with the solenoid only* are noticeably constrained a bit more when looking at the screen vs. off (at least on dual core mode). The values therein are optimized for when the gun sees IR.