Added simple player numbers
New define added that sets the player number, which adjusts Start/Select to the necessary default buttons unique to that slot automatically without tampering with the further. Only use 1-4, or you'll get an error!

Note: should we maybe be using a gamepad device for the cursor keys? That could be neat, and mean that the d-pad->arrow keys won't be shared between all of them... but then again, who or what the heck multiplayer games uses the d-pad?

Clarified bits of documentation while I'm at it.