Implement Serial Control commands
In the case that you have a gun that does not allow any/convenient access to Pause Mode and all the functions therein, this will expose a couple of quick shortcuts to certain modes and settings.

All of these are prefixed with 'X' to distinguish from normal MAMEHOOKER functions (but also requires using that mode atm). As follows:
'XP' toggles Pause Mode
'XT' toggles Processing Sketch-compatible run mode
'XC' immediately sets the gun to calibrate (add a number to optionally switch to that profile immediately prior to calib)
'XS' saves current preferences to EEPROM/Flash
'X' followed by a number from 1-4 switches the gun to any one of the four available calibration profiles

Also added the ability for pause mode to read Serial commands. This might cause havoc if Mamehook is (about to be) active, so it might require some guardrails and sanity checks if this becomes a more serious issue.