Finish/add Gamepad output support
I'm guessing this was intended to be done, given how the ReportTypes struct had a label for Gamepad commented out - but since we're adding Gamepads output support, we can now make use of that.

Also moved the offScreen flag to be local to LightgunButtons, so the library can now distinguish on/offscreen reports (though it's still updated by the sketch, of course). The new analogOutput flag is local to this library too for similar reasons.

Finishing the fight.