Mega Update pt. 1
Pt. 1 of this update adds a third fully functional Gamepad output mode.
Currently only enable-able from the Serial (using the new code 'XA'), using this will toggle between the current default Mouse output to a gamepad with 16 bits of analog precision - 65535 maximum positions on both X and Y axis, which should be more than enough. 

This is intended to be used mainly for the current PCSX2 nightly, or other games that treat the left stick of a controller as a surrogate absolute positioning device. There shouldn't be any noticeable difference in performance from the gun itself, aside from how the game/emulator itself in question handles controller latency.

This update also expands the Button Descriptor array, now allowing for dictating different functions for buttons depending on if the gun is pointed on or offscreen! Note, that this is distinct from the pre-existing "Offscreen Button Mode" toggle for buggy games, although it is similar in functionality. This is mainly intended for guns with limited buttons, like the Namco GunCon 1, Konami Justifier, or Sega Virtua Gun/Saturn Stunner.

The offScreen flag is now a part of the LightgunButtons library, so changed references to it to reflect this change.

Also fixes part of the TinyUSB_Devices init, since the library has changed to accommodate Gamepad reporting and using an appropriate name for the begin function.

And other small changes that are meant to be reflected with pt. 2...