Added simple pause menu and hold-to-pause functionality
Another feature intended for guns with limited buttons, this adds a simple pause menu with an interface that the user scrolls through options using the A &/or B buttons, and activates a selected option with the Trigger. The menu loops back around to the end/start when scrolling beyond when at the first and last options, hides feedback activation options when switches are selected, can be canceled with the C Button/Reload, and will always start with Calibrating the current profile as the initial menu selection.

Menu options are as follows:
 - Calibrate current profile
 - Switch to a new profile (terminates in a submenu where you can select any of the four profiles, starting from your currently selected one)
 - Save settings to EEPROM/Flash
 - Toggle Rumble
 - Toggle Solenoid
 - Toggle Burst Firing mode
 - Exit the menu and return to the normal run mode

As of now, the onboard LED (if available) doesn't change outside of profile colors when scrolling through options in the profile selection submenu. Current selections are outputted in the serial monitor if connected. This simple menu option is disabled by default, but can be enabled by setting `simpleMenu` to true.

Also added is the ability to enter pause mode (either simple pause menu or the original hotkey-based pause mode) by holding the assigned button combo EnterPauseModeHoldBtnMask (set by default to Trigger + A Button) only while aiming off-screen. Pressing any buttons more or less than the assigned combo, OR aiming on-screen, will cancel this out.
This setting is not enabled by default, and can be set on with `holdToPause`. The amount of time that the combo needs to be held is configurable under `pauseHoldLength`, as milliseconds.