Calibration is one-button compatible now
If you're in the (un)fortunate position of using a Sega Stunner/Virtua Gun shell without additional buttons, the horizontal & vertical calibration steps can now be adjusted with the single A button - pressing it (A) while aiming on-screen will decrease the scale, and pressing (A) while the cursor is at the edge of the display will increase the scale; essentially serving the same purpose as pressing the B button.

As a side effect of this implem, pressing B now takes priority over A in calibration - the old code would simply cancel out the scale change if both buttons were pressed. If you're using anything more reasonable like a normal GunCon, then this won't change anything in normal use.

It's not perfect, but the Virtua Gun isn't either, so it all works out in the end.