Only check & process serial once every cycle
I've found that, at least on Linux, Mamehooker can get into a frenzy and flood the serial line so much that no inputs are received (which would block buttons inputs in dual core, and outright stop everything in single core mode).
Since we're fast enough as it is in dual core, only process serial once if it's detected--should Mamehook do something catastrophic(ally stupid) like that, the gun can still at least be used. Also minimized outputs for commands that are expected to be used by Mamehook, in case the backtalk is what's causing confusion.