add placeholders for pump action and home button pins
The Top Shot Elite has by far the most buttons in a vanilla light gun, so two more spaces are added to accommodate this behemoth;
Pump action, which maps to right mouse, since it's commonly used as a reload, and the Home button, which is used as a Pause Mode button.
Still deciding on default pins for these...

Oh, and pins are now signed byte variables instead of unsigned constants. This will facilitate being able to selectively exclude or skip buttons mapped to pins that are not using a valid number (for now, anything that's negative). Will be important when save/load functionality is expanded.