Support external NeoPixels, unify/cleanup LED operations
Part of this involves moving to the Adafruit NeoPXL8 library, which not only is more efficient than the original reference lib, but also happens to fit very conveniently with GUN4ALL's method of reserving a space for an external LED if enabled! It fits like a glowy glove.

Enable by uncommenting `CUSTOM_NEOPIXEL`. If it's undefined (negative numbers), the library will handle it fine regardless, so it'll still work when we implement pins layout saving. Note that we only use the first "pixel" of any given neopixel pin--I'm not sure how many, if any, people will use strips; if you do, please reach out!

Cleaned up LEDs usage by unifying all of what were originally pointers and preprocessor defines for the individual methods into one `LedUpdate()` function, which does the same thing but commits to all detected devices. Also added a sanity check for the fourpin LED, which just skips the functionality if the feature is enabled but the pins are invalid.

All that's left is translating Wikicolors to three-color channels for fourpins, which would then complete the RGB implementation.