Fixed brokenness caused by PXL8
Okay so NeoPXL8 doesn't seem to work, for whatever reason--so rolled back to the stock NeoPixel library as it's the only one flexible enough and working for our needs.

Unfortunately, PXL8 clamped all of our devices to a single ascending index which would have been helpful, so using dynamically allocated pins/enabling required a different approach, but worked out a bit better in the end;

The NeoPixel object will always be created at start (it seems to store in an unsigned var so that's potentially dangerous), but the external Pixel will only be initialized and accessed if the pin number is valid. All calls to RGB--and thus to NeoPixels--will also check this number to make sure it's correct.

On the plus side, the strip length is now configurable and will be updated all at once for both packed and separated color values, so this should futureproof things a bit. Just don't throw on too big a strip, for the sake of performance.