Extract indiv. color channels from packed color value for fourpins
Turns out that the `color` value is packed in a way where the Blue value is the least significant byte, Green is the second-least significant, and Red is the third-least significant (with the most significant byte I THINK reserved for the White channel, for RGBW NeoPixels). So, we extract these bytes into the typical R/G/B values and just hope they're close enough to what Prow intended these to look.

Red is fun, since it's out of the typical scope of low/highByte (as it only covers the least-two significant bytes of a long), so we shift the value by a byte to grab the Red channel.

So at least now, Fourpins should be fully supported in all usecases.