Analog Stick input compatibility
Including validation checking (to make sure pins don't clash with the temp sensor or to ignore when it's enabled but unset), sane opportunity checking to mitigate overloading USB bandwidth, and a very basic deadzone implementation. The gun stick is mapped to Left, with analog output mode set to the right.

Also adjusted the escape key signal so that it should be recognized more consistently.

Haven't actually tested, but in *theory* should work fine. Though I seem to have run into stability issues in pre-preliminary testing when the analog polling isn't set in the right place...