Attempt no. 527 at serial crash mitigations
We surround the Serial processing step with 2ms pauses on core 2, and sends a signal for core 1 to pause as well either prior to or after updating the camera position. No noticeable degradation in performance as far as I can tell, and seems to at least prevent instant crashes.
Pauses only happen if serial data is seen in gun run mode.