Removed manual burst fire toggle
Since the (few) games that use/benefit from an always-on three-shot burst fire are *probably* supported by Mamehooker anyways, there's no need to have a button combo dedicated to it. It wasn't ever planned to be implemented in simple menu anyways, so better to remove it for simplicity. This will still be enabled if set by the serial toggle `M8.1`.

While we're at it, clean up the actual "demo" functionality of the toggles because wow they were done amateurishly.