Added extended save/load support
Now every setting in the sketch can be set at runtime over serial commands (see the lowercase entries under 'X'-type serial commands in SerialProcessing).

Pin numbers can now be set at boot time, rather than hardcoded - currently these values only update at boot, due to the dependence on all the pin modes being set in the setup process. Pin assignments will always fallback to the values hardcoded in the sketch unless `customPinsInUse` is enabled and saved thereafter; custom pins can then be set and saved on the next boot.
Committing extended preferences to EEPROM with custom pins setting disabled will overwrite the values in saveram back to the defaults as set in the sketch.

All of the adjustables are now unsigned integers for the sake of save simplicity, though variables that were originally bytes are clamped to the original 0-255 range when attempting to reassign them with `X.m.2.x.y` so it should still be safe.