Low button mode formalized, fix pins loading
LowButtonMode is a shortcut to remapping the A & B buttons as Start & Select when aiming offscreen. Set with `Xm.0.6.x` - changes are reflected immediately, and applied at startup if extended prefs are saved (`Xs`).

This also fixes custom pins mapping, re-committing the changes to the buttons input table properly so input should work with the custom pins mapping flag enabled.

Since we now need to load the buttons in the main core, the second core will go asleepy now for the boot process.