Safety: Constrain all setting inputs
Less necessary for the bools, but this ensures Pin numbers are mapped within a reasonable range.

Also adds more helpful labels if a bool is toggled ON or OFF.

For anyone that cares, the pin range numbers are pure pulled-out-of-my-ass-jutsu. If your microcontroller somehow has this many accessible GPIO ports, let alone more, then I wanna see pics.