Add docked mode, irsens and runmode setting
The `XP` pause command is now used to set the gun into "Docked Mode" - when invoked by the GUI app (soon[tm]), it allows for safe save operation and will send bits whenever a button is pressed and released. Currently, the only way this mode can be accessed is from the GUI (but can be access from first boot, if the gun is prompting for input).

Misc changes made to facilitate docked mode.
Removed needless printout when typing `Xm`, since the "setting" commands are now exclusively for GUI use.
Board can now print out saved profile table for GUI use, and remotely set IR sensitivity and run mode averaging.

Believe it or not, the initial GUI is very close to being done! So this should (HOPEFULLY) be the last GUI-related update to the firmware directly, at least for a while.