Split off serial processing for docked usage
The dangerous stuff is now parted out to the SerialProcessingDocked method, only accessible when the gun's in Docked mode (or in the first-time setup UX). Should prevent potentially ugly scenarios, like if MAMEHOOKER take a dump. Also isn't lumped in with Mamehooker build flag (though the sketch will still not build without it).

Changed test printouts so that it's blocked by committing ops in SerialProcessingDocked, and unlocked after a save.

todo note for update bindings method, because it continues to make me cringe.