Pico W & OLED support
And the wrapper lib is actually finished.
This should be about as good as the OLED could functionally be, with the only known issue being that some profile names in the simple pause profile list might be slightly corrupted for whatever reason? I can only chalk it up to a bug in the display though, as logically it shouldn't be an issue.

Also adds Ammo and Life counts from Mamehook et al for the display to use.