Life Bar support, mid-serial pause fixes
Life Bar is supported whenever `B` is appended to a Display Modeset call. It displays the current life (out of an assumed percentile of 100) both graphically and in subtext. Bars scale appropriate when in single element or dual element display modes.
Solo Life Glyphs are now bigger to use more of the screen real estate when it's the only thing shown.
No more serial reading in Pause Mode, as it has too much potential to break shit (it was only really used back when things were set by manually typing in commands to a serial monitor, but there's no need for it now).
Display wrapper lib is the one aware of the current screen state, and better enforces the settings it was configured with at serial handoff init time.
Removed the `lastAmmo/lastLife` system, as it doesn't really optimize much of anything in practical use - an equivalent `currentAmmo/currentLife` is now used for simply redrawing the display in serial handoff mode from non-runmode states.