Main: Cleanup main by splitting up INO definitions into either dedicated classes or a header for the main file.
With this, all definitions are normal C-friendly, and the main INO becomes significantly less cluttered. Serial and shared variables are now in their own classes for organization/code clarity.
Also integrates with the new shared constants file, structs are converted to fixed-size arrays, default mappings are set via the shared defaults.
As a bonus, cali mode is initiated directly from when it's called, and will skip sending mouse movement when called from the Desktop App.
And a couple of other small tweaks that I'm forgetting atm...