Update OpenFIREFeedback.cpp
The rumble pin is used in analog PWM mode, therefore analogWrite must be used and not digitalWrite and even though with the library for rp2040 it still works when the low level of the pin is set (with this library it is equivalent to writing analogWrite(pin, 0)), with other micros such as esp32 it does not work and you must use analogWrite setting the value to 0. Therefore I ask the courtesy, since I have adapted OpenFire to also work with esp32, to make this modification, which does not involve anything different in the operation. I have tested it both on rp2040 and on esp32 and with analogWrite it works well with both micros.