Update SamcoPreferences.cpp
The EEPROM.update method does not exist in other EEPROM libraries such as the one for the esp32, while the EEPROM.put method exists for all libraries. The EEPROM.update method in rp2040 does nothing but call the EEPROM.put method and nothing more. I have adapted OpenFIRE to work also on EPS32 and if there are no problems, I ask the courtesy of making this small change that does not affect the operation.