Update SamcoEnhanced.ino
Removed 'TUSBDeviceSetup' and put back 'TinyUSBDevices', or did you just start the replacement and want to remove all the 'TinyUSBDevices' to replace them with the new 'TUSBDeviceSetup'? .. otherwise, I don't see the point, but maybe it's me who can't understand it because I'm an idiot. It also works with 'TUSBDeviceSetup' but put like this it doesn't make sense, at least I think so